--- advent_of_code/aoc/2022/inputs/day17_sample.txt
@@ -1,1 +1,0 @@
->>><<><>><<<>><>>><<<>>><<<><<<>><>><<>>
